Split schedule and bots into separate microservices
