+++ app_schedule_telegram_bot/Cargo.toml
@@ -6,3 +6,11 @@ authors = ["Anton Kolomeytsev <tonykolomeytsev@gmail.com>"]
 readme = "README.md"
 
 [dependencies]
+common_actix = { workspace = true }
+common_errors = { workspace = true }
+domain_schedule_models = { workspace = true }
+domain_telegram_bot = { workspace = true }
+feature_telegram_bot = { workspace = true }
+
+actix-web = { workspace = true }
+anyhow = { workspace = true }
